Moved espresso tests into LottieSample (#1264)
