+++ LottieSample/src/androidTest/java/com/airbnb/lottie/samples/ObjectPool.kt
@@ -1,9 +1,8 @@
-package com.airbnb.lottie
+package com.airbnb.lottie.samples
 
-import android.graphics.Bitmap
 import android.util.Log
+import com.airbnb.lottie.L
 import java.util.*
-import java.util.concurrent.Semaphore
 import kotlin.collections.HashSet
 
 internal class ObjectPool<T>(private val factory: () -> T) {
