+++ LottieSample/src/androidTest/java/com/airbnb/lottie/samples/SuspendingSemaphore.kt
@@ -1,4 +1,4 @@
-package com.airbnb.lottie
+package com.airbnb.lottie.samples
 
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.runBlocking
