+++ LottieSample/src/main/AndroidManifest.xml
@@ -73,6 +73,7 @@
             android:exported="true" />
 
         <activity android:name=".SnapshotTestActivity" />
+        <activity android:name=".EmptyActivity" />
         <service android:name="com.amazonaws.mobileconnectors.s3.transferutility.TransferService" android:enabled="true" />
     </application>
