+++ LottieSample/src/main/res/layout/activity_empty.xml
