+++ LottieSample/src/main/res/layout/auto_play.xml
