+++ LottieSample/src/main/res/layout/auto_play_gone.xml
