+++ LottieSample/src/main/res/layout/film_strip_view.xml
@@ -6,127 +6,127 @@
     android:layout_gravity="center"
     android:columnCount="5">
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_1"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_2"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_3"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_4"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_5"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_6"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_7"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_8"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_9"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_10"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_11"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_12"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_13"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_14"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_15"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_16"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_17"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_18"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_19"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_20"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_21"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_22"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_23"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_24"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
 
-    <com.airbnb.lottie.NoCacheLottieAnimationView
+    <com.airbnb.lottie.samples.NoCacheLottieAnimationView
         android:id="@+id/animation_25"
         android:layout_width="@dimen/film_strip_size"
         android:layout_height="@dimen/film_strip_size" />
