+++ LottieSample/src/main/res/layout/no_auto_play.xml
