+++ LottieSample/src/main/res/raw/lottielogo.json
