--- espresso/build.gradle
@@ -1,45 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
-android {
-    compileSdkVersion 28
-
-
-    defaultConfig {
-        applicationId "com.airbnb.lottie.tests.espresso"
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-        }
-    }
-
-}
-
-dependencies {
-    implementation project(':lottie')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
-    implementation "androidx.recyclerview:recyclerview:$androidXVersion"
-    implementation 'androidx.appcompat:appcompat:1.1.0-alpha05'
-    implementation 'androidx.fragment:fragment:1.1.0-alpha09'
-    implementation 'androidx.core:core:1.2.0-alpha01'
-    implementation 'androidx.core:core-ktx:1.1.0-beta01'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    implementation 'androidx.test.espresso:espresso-idling-resource:3.2.0'
-    debugImplementation("androidx.fragment:fragment-testing:1.1.0-alpha09") {
-        exclude group: 'androidx.test'
-    }
-    androidTestImplementation 'androidx.test.ext:junit:1.1.0'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
-    androidTestImplementation 'androidx.fragment:fragment-testing:1.1.0-alpha09'
-    androidTestImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:2.1.0"
-    androidTestImplementation "org.mockito:mockito-android:2.28.2"
-}
