+++ lottie/build.gradle
@@ -26,7 +26,7 @@ android {
 }
 
 dependencies {
-  implementation "androidx.appcompat:appcompat:${androidXVersion}"
+  implementation "androidx.appcompat:appcompat:1.0.0"
   testImplementation "org.mockito:mockito-core:2.15.0"
   testImplementation 'junit:junit:4.12'
   testImplementation "org.robolectric:robolectric:4.0-alpha-3"
